chore: moved monomorphization tests
