+++ tests/monomorphizer.rs
@@ -1,0 +1,27 @@
+use slynx_hir::{SlynxHir, error::HIRErrorKind};
+use slynx_lexer::Lexer;
+use slynx_parser::Parser;
+use slynx_monomorphizer::Monomorphizer;
+use slynx_typechecker::TypeChecker;
+
+#[test]
+fn rejects_cyclic_aliases_in_monomorphization() {
+    let source = "alias A = B; alias B = A; func main(): void {}";
+    let tokens = Lexer::tokenize(source).expect("source should tokenize");
+    let declarations = Parser::new(tokens)
+        .parse_declarations()
+        .expect("source should parse");
+    let mut hir = SlynxHir::new();
+    hir.generate(&declarations).expect("HIR should generate");
+
+    let mut types_module = TypeChecker::check(&mut hir).expect("type checking should pass");
+
+    let err =
+        Monomorphizer::resolve(&hir, &mut types_module)
+            .expect_err("cyclic aliases should fail during monomorphization");
+
+    match &err.kind {
+        HIRErrorKind::RecursiveType { ty } => assert_eq!(*ty, hir.modules.intern_name("A")),
+        other => panic!("expected RecursiveType, got {other:?}"),
+    }
+}
